2D conZono plotting update
Fixed a bug that causes vertices of the set to be plotted out of order, which is important when using the patch command in MATLAB.  The new method still requires setting a tolerance when determining if a vertex has already been found or not. There could still be issues depending on how this tolerance relates to the tolerance of the linear program solver used to find the vertices.